Split OAuth2GrantType and OAuth2GrantTypeFactory
