--- services/src/main/resources/META-INF/services/org.keycloak.protocol.oidc.grants.OAuth2GrantType
@@ -1,9 +1,0 @@
-org.keycloak.protocol.oidc.grants.AuthorizationCodeGrantType
-org.keycloak.protocol.oidc.grants.ClientCredentialsGrantType
-org.keycloak.protocol.oidc.grants.PermissionGrantType
-org.keycloak.protocol.oidc.grants.RefreshTokenGrantType
-org.keycloak.protocol.oidc.grants.ResourceOwnerPasswordCredentialsGrantType
-org.keycloak.protocol.oidc.grants.TokenExchangeGrantType
-org.keycloak.protocol.oidc.grants.ciba.CibaGrantType
-org.keycloak.protocol.oidc.grants.device.DeviceGrantType
-
